backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
qdrant-client>=1.7.0
google-generativeai>=0.4.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
cohere>=4.0.0
beautifulsoup4>=4.12.0
requests>=2.30.0
slowapi>=0.1.8